refactor, add missed propTypes
