+++ src/containers/FilmList.js
@@ -1,5 +1,5 @@
 import React from "react";
-import FilmElement from "./FilmElement";
+import FilmElement from "../components/FilmElement";
 import PropTypes from 'prop-types';
 import {withStyles} from '@material-ui/core/styles';
 import GridList from '@material-ui/core/GridList';
@@ -14,16 +14,13 @@ const styles = theme => ({
     overflow: 'hidden',
     backgroundColor: theme.palette.background.paper,
   },
-  gridList: {
-    // width: 1000,
-  },
 });
 
 const FilmList = props => {
   const {filmList, selectFilm, classes} = props;
   return (
     <div className={classes.root}>
-      <GridList cellHeight={180} className={classes.gridList}>
+      <GridList>
         {filmList.map(film =>
           <FilmElement
             key={film.imdbID}
@@ -38,6 +35,7 @@ const FilmList = props => {
 FilmList.propTypes = {
   classes: PropTypes.object.isRequired,
   filmList: PropTypes.array.isRequired,
+  selectFilm: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = state => ({
