+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './containers/App';
+import App from './components/App';
 import {Provider} from "react-redux";
 import {createStore} from "redux";
 import reducer from "./reducers";
